requirements.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.3.0
altair==5.5.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dlib==19.24.6
efficientnet-pytorch==0.7.1
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.0
idna==3.10
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.36.1
imgaug==0.4.0
imutils==0.5.4
Jinja2==3.1.4
joblib==1.4.2
jpeg4py==0.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
mdurl==0.1.2
narwhals==1.16.0
networkx==3.4.2
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==24.2
pandas==1.5.3
pillow==11.0.0
protobuf==5.29.0
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qudida==0.0.4
referencing==0.35.1
requests==2.32.3
retinaface-pytorch==0.0.7
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
smmap==5.0.1
streamlit==1.40.2
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.9.20
toml==0.10.2
torch==1.13.1
torch-dct==0.1.6
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.3
